\name{chooseBreaks-methods}
\docType{methods}
\alias{chooseBreaks-methods}
\alias{chooseBreaks,list,numeric-method}
\title{chooseBreaks}
\description{
Choose evenly spaced breakpoints from a list of breakpoints.
}
\section{Methods}{
\describe{

\item{\code{signature(breakpoints="list",numbreaks="numeric")}}{
Returns a sublist of breakpoints of length \code{numbreaks}.
}
}}
\author{
Diana H.P. Low, Efthimios Motakis
}
\keyword{methods}

