\name{clusterPV-methods}
\docType{methods}
\alias{clusterPV-methods}
\alias{clusterPV,SegTrajectories-method}
\title{clusterPV}
\description{
Returns a pvclust values for object of class "SegTrajectories")
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SegSeriesTrajectories")}}{
Returns an object of class "pvclust".
}

}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}
\seealso{
\code{\link{transformSeries}}
}
\examples{
#data(deltaGseg)
#clusterPV(td2)
}
\keyword{methods}
