\name{clusterSegments-methods}
\docType{methods}
\alias{clusterSegments-methods}
\alias{clusterSegments,SegTrajectories-method}
\title{clusterSegments}
\description{
Performs the function \code{clusterSegments} on an object of class "SegTrajectories".
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SegTrajectories")}}{
Returns an object of class "SegSeriesTrajectories"
}

}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}

\examples{
#data(deltaGseg)
#st<-clusterSegments(dt)
}
\keyword{methods}
