\name{deltaGseg}
\alias{deltaGseg}
\docType{package}
\title{
deltaGseg
}
\description{
deltaGseg
}
\details{
\tabular{ll}{
Package: \tab deltaGseg\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2013-02-04\cr
License: \tab GPL-2\cr
}
}
\author{
Diana H.P. Low, Efthimios Motakis

Maintainer: Diana H.P. Low <dlow@imcb.a-star.edu.sg>
}
\references{
Zhou W., Motakis E., Fuentes G., Verma C.S. (2012) Macrostate identification from biomolecular simulations through time series analysis. J Chem Inf Model. 2012 Sep 24;52(9):2319-24. Epub 2012 Sep 5.

Dickey, D.A. and W.A. Fuller (1979), Distribution of the Estimators for Autoregressive Time Series with a Unit Root. Journal of the American Statistical Association, 74, p. 427-431.

Nason, G.P. (2008) Wavelet methods in Statistics with R. Springer, New York.

Auger, I. E.; Lawrence, C. E. Algorithms for the optimal identification of segment neighborhoods. Bull. Math.Biol. 1989, 51(1), 39-54.

Shimodaira, H. (2004) Approximately unbiased tests of regions using multistep-multiscale bootstrap resampling, Annals of Statistics, 32, 2616-2641.

D'Agostino R.B., Pearson E.S. (1973); Tests for Departure from Normality, Biometrika 60, 613-22.

}
\keyword{ package }
