\name{diagnosticPlots-methods}
\docType{methods}
\alias{diagnosticPlots-methods}
\alias{diagnosticPlots,SegSeriesTrajectories-method}
\title{diagnosticPlots}
\description{
Performs the function \code{diagnosticPlots} on an object of class "SegSeriesTrajectories".
}
\section{Methods}{
\describe{
\item{\code{signature(object = "SegSeriesTrajectories")}}{
Returns histogram and acf plots of series
}
}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}
\examples{
#data(deltaGseg)
#diagnosticPlots(st)
}
\keyword{methods}
