\name{getIntervals-methods}
\docType{methods}
\alias{getIntervals-methods}
\alias{getIntervals,SegSeriesTrajectories-method}
\title{getIntervals}
\description{
Returns a list of subpopulations and the intervals of the segmented series.)
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SegSeriesTrajectories")}}{
Returns a list of subpopulations and the intervals of the segmented series from \code{segmentationWithinSeries}
}

}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}
\seealso{
\code{\link{transformSeries}}
}
\examples{
#data(deltaGseg)
#getIntervals(traj1.ss)
}
\keyword{methods}
