\name{getTNames-methods}
\docType{methods}
\alias{getTNames-methods}
\alias{getTNames,Trajectories-method}
\alias{getTNames,TransTrajectories-method}
\title{Retrieve filenames of trajectories}
\description{
Returns a vector from the @filenames or @tfilenames slot.
}
\details{
Returns a vector of filename(s) values.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Trajectories")}}{
Returns the original filenames of series read by parseTraj.
}

\item{\code{signature(object = "TransTrajectories")}}{
Returns filenames generated by the transformSeries function if the originial series has been split into subseries. Subseries names are denoted by an underscore after the original names, eg. FILE1_1, FILE1_2.
}
}}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}

\examples{
#data(deltaGseg)
#getTNames(traj1)
}

\keyword{methods}
