\name{getTraj-methods}
\docType{methods}
\alias{getTraj-methods}
\alias{getTraj,Trajectories-method}
\alias{getTraj,TransTrajectories-method}
\title{ getTrajectories}
\description{
Returns a list of matrices from the @trajlist slot (for object of class "Trajectories") or @ttrajlist slot (for object of class "TransTrajectories")
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Trajectories")}}{
Returns the original trajectories used for computation.
}

\item{\code{signature(object = "TransTrajectories")}}{
Returns the transformed trajectories generated by the function \code{\link{transformSeries}}
}
}}
\author{
Diana H.P. Low, Efthimios Motakis
}
\keyword{methods}

