\name{plot-methods}
\docType{methods}
\alias{plot}
\alias{plot-methods}
\alias{plot,SegSeriesTrajectories-method}
\alias{plot,SegTrajectories-method}
\alias{plot,Trajectories-method}
\alias{plot,TransTrajectories-method}
\title{Plot Trajectories-related objects}
\description{
Plot "Trajectories" objects and customizes output.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "Trajectories")}}{
  \code{plot(object,name='all',breakpoints=NULL)}\cr\cr
  \code{name}: Character. Name of sub-series, or if \code{all}, plots the whole series.\cr
  \code{breakpoints}: List. Supply breakpoints generated by \code{\link{splitTraj}}.
}

\item{\code{signature(x = "TransTrajectories")}}{
  \code{plot(object,labelling=TRUE)}\cr\cr
  \code{labelling}: Logical. Writes labels. May be turned off to prevent overcrowding of plot.
}
\item{\code{signature(x = "SegTrajectories")}}{
  \code{plot(object)}
}
\item{\code{signature(x = "SegSeriesTrajectories")}}{
  \code{plot(object)}
}
}}
\keyword{methods}
