\name{plotDiff-methods}
\docType{methods}
\alias{plotDiff-methods}
\alias{plotDiff,TransTrajectories-method}
\title{plotDiff}
\description{
Plots differentiated (sub)series in object of class "TransTrajectories")
}
\section{Methods}{
\describe{

\item{\code{signature(object = "TransTrajectories")}}{
Plots (sub)series before and after transformation.
}

}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}
\seealso{
\code{\link{transformSeries}}
}
\examples{
#data(deltaGseg)
#plotDiff(traj1.tr,'D_GBTOT3_1')
}
\keyword{methods}
