\name{splitTraj-methods}
\docType{methods}
\alias{splitTraj-methods}
\alias{splitTraj,Trajectories-method}
\title{splitTraj}
\description{
Performs the function \code{splitTraj}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Trajectories")}}{
Returns a list of breakpoints identified in the trajectory series.
}

}}
\keyword{methods}
\references{
None
}
\author{
Diana H.P. Low, Efthimios Motakis
}

\examples{
#data(deltaGseg)
#all_breakpoints<-splitTraj(traj1)
}
\keyword{methods}
